+++ backend/src/main/java/com/example/backend/Controllers/PayController.java
@@ -1,0 +1,34 @@
+package com.example.backend.Controllers;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.example.backend.Entities.Payment;
+import com.example.backend.Services.PaymentServices;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("payment")
+public class PayController {
+
+
+    private final PaymentServices paymentServices;
+
+
+    @PostMapping
+    public ResponseEntity<String> pay(@RequestBody Payment payment) throws Exception {
+        try {
+            this.paymentServices.pay(payment);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+}
